+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove SecretReconciler.
+- Refactor Reconcilers.
+
 ### Added
 
 - Add new role for AWS Load Balancer Controller.
+- Add tests for iam package.
 
 ## [0.10.0] - 2023-08-11