+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-10-27
+
 ### Added
 
 - Add `global.podSecurityStandards.enforced` value for PSS migration.
@@ -140,7 +142,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement `AWSMachineTemplate` reconciler.
 - Implement `AWSMachinePool` reconciler.
 
-[Unreleased]: https://github.com/giantswarm/capa-iam-operator/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/giantswarm/capa-iam-operator/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/giantswarm/capa-iam-operator/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/giantswarm/capa-iam-operator/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/giantswarm/capa-iam-operator/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/giantswarm/capa-iam-operator/compare/v0.7.0...v0.8.0